+++ django_filters/fields.py
@@ -5,13 +5,11 @@ from datetime import datetime, time
 from collections import namedtuple
 
 from django import forms
+from django.conf import settings
 from django.utils.dateparse import parse_datetime
+from django.utils import timezone
 
-# TODO: Remove this once Django 1.4 is EOL.
-try:
-    from django.utils.encoding import force_str
-except ImportError:
-    force_str = None
+from django.utils.encoding import force_str
 
 from .widgets import RangeWidget, LookupTypeWidget
 
@@ -61,6 +59,8 @@ class TimeRangeField(RangeField):
 
 
 Lookup = namedtuple('Lookup', ('value', 'lookup_type'))
+
+
 class LookupTypeField(forms.MultiValueField):
     def __init__(self, field, lookup_choices, *args, **kwargs):
         fields = (
@@ -75,7 +75,7 @@ class LookupTypeField(forms.MultiValueField):
         super(LookupTypeField, self).__init__(fields, *args, **kwargs)
 
     def compress(self, data_list):
-        if len(data_list)==2:
+        if len(data_list) == 2:
             return Lookup(value=data_list[0], lookup_type=data_list[1] or 'exact')
         return Lookup(value=None, lookup_type='exact')
 
@@ -91,15 +91,21 @@ class IsoDateTimeField(forms.DateTimeField):
     """
     ISO_8601 = 'iso-8601'
     input_formats = [ISO_8601]
+    default_timezone = timezone.get_default_timezone() if settings.USE_TZ else None
 
     def strptime(self, value, format):
-        # TODO: Remove this once Django 1.4 is EOL.
-        if force_str is not None:
-            value = force_str(value)
+        value = force_str(value)
 
         if format == self.ISO_8601:
             parsed = parse_datetime(value)
             if parsed is None:  # Continue with other formats if doesn't match
                 raise ValueError
+
+            # Handle timezone awareness. Copied from:
+            # https://github.com/tomchristie/django-rest-framework/blob/3.2.0/rest_framework/fields.py#L965-L969
+            if settings.USE_TZ and not timezone.is_aware(parsed):
+                return timezone.make_aware(parsed, self.default_timezone)
+            elif not settings.USE_TZ and timezone.is_aware(parsed):
+                return timezone.make_naive(parsed, timezone.UTC())
             return parsed
         return super(IsoDateTimeField, self).strptime(value, format)