+++ runshell.py
@@ -1,29 +1,12 @@
 #!/usr/bin/env python
+import os
 import sys
-from django.conf import settings
-from django.core.management import call_command
 from django.core.management import execute_from_command_line
-
-if not settings.configured:
-    settings.configure(
-        DATABASES={
-            'default': {
-                'ENGINE': 'django.db.backends.sqlite3',
-                'NAME': ':memory:',
-            },
-        },
-        INSTALLED_APPS=(
-            'django_filters',
-            'tests',
-        ),
-        ROOT_URLCONF=None,
-        USE_TZ=True,
-        SECRET_KEY='foobar'
-    )
 
 
 def runshell():
-    call_command('syncdb', interactive=False)
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "tests.settings")
+    execute_from_command_line(sys.argv[:1] + ['migrate', '--noinput', '-v', '0'])
     argv = sys.argv[:1] + ['shell'] + sys.argv[1:]
     execute_from_command_line(argv)