+++ runtests.py
@@ -1,37 +1,11 @@
 #!/usr/bin/env python
+import os
 import sys
-from django import VERSION
-from django.conf import settings
 from django.core.management import execute_from_command_line
-
-if not settings.configured:
-    test_runners_args = {}
-    if VERSION < (1, 6):
-        test_runners_args = {
-            'TEST_RUNNER': 'discover_runner.DiscoverRunner',
-        }
-    settings.configure(
-        DATABASES={
-            'default': {
-                'ENGINE': 'django.db.backends.sqlite3',
-                'NAME': ':memory:',
-            },
-        },
-        INSTALLED_APPS=(
-            'django.contrib.contenttypes',
-            'django.contrib.auth',
-            'django_filters',
-            'tests',
-        ),
-        ROOT_URLCONF=None,
-        USE_TZ=True,
-        SECRET_KEY='foobar',
-        SILENCED_SYSTEM_CHECKS=['1_7.W001'],
-        **test_runners_args
-    )
 
 
 def runtests():
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "tests.settings")
     argv = sys.argv[:1] + ['test'] + sys.argv[1:]
     execute_from_command_line(argv)