+++ tests/test_filterset.py
@@ -1,13 +1,9 @@
 from __future__ import absolute_import, unicode_literals
 
 import mock
-import sys
+import unittest
 
-if sys.version_info >= (2, 7):
-    import unittest
-else:  # pragma: nocover
-    from django.utils import unittest  # noqa
-
+import django
 from django.db import models
 from django.test import TestCase
 
@@ -19,6 +15,7 @@ from django_filters.filters import NumberFilter
 from django_filters.filters import ChoiceFilter
 from django_filters.filters import ModelChoiceFilter
 from django_filters.filters import ModelMultipleChoiceFilter
+from django_filters.filters import UUIDFilter
 
 from .models import User
 from .models import AdminUser
@@ -35,6 +32,7 @@ from .models import DirectedNode
 from .models import Worker
 from .models import HiredWorker
 from .models import Business
+from .models import UUIDTestModel
 
 
 def checkItemsEqual(L1, L2):
@@ -42,6 +40,7 @@ def checkItemsEqual(L1, L2):
     TestCase.assertItemsEqual() is not available in Python 2.6.
     """
     return len(L1) == len(L2) and sorted(L1) == sorted(L2)
+
 
 class HelperMethodsTests(TestCase):
 
@@ -80,7 +79,7 @@ class DbFieldDefaultFiltersTests(TestCase):
             models.FilePathField,
             models.FloatField,
             models.IntegerField,
-            models.IPAddressField,
+            models.GenericIPAddressField,
             models.NullBooleanField,
             models.PositiveIntegerField,
             models.PositiveSmallIntegerField,
@@ -92,6 +91,7 @@ class DbFieldDefaultFiltersTests(TestCase):
             models.ForeignKey,
             models.OneToOneField,
             models.ManyToManyField,
+            models.UUIDField,
         ]
         msg = "%s expected to be found in FILTER_FOR_DBFIELD_DEFAULTS"
 
@@ -102,7 +102,6 @@ class DbFieldDefaultFiltersTests(TestCase):
         to_check = [
             models.Field,
             models.BigIntegerField,
-            models.GenericIPAddressField,
             models.FileField,
             models.ImageField,
         ]
@@ -119,6 +118,12 @@ class FilterSetFilterForFieldTests(TestCase):
         result = FilterSet.filter_for_field(f, 'username')
         self.assertIsInstance(result, CharFilter)
         self.assertEqual(result.name, 'username')
+
+    def test_filter_found_for_uuidfield(self):
+        f = UUIDTestModel._meta.get_field('uuid')
+        result = FilterSet.filter_for_field(f, 'uuid')
+        self.assertIsInstance(result, UUIDFilter)
+        self.assertEqual(result.name, 'uuid')
 
     def test_filter_found_for_autofield(self):
         f = User._meta.get_field('id')
@@ -183,7 +188,7 @@ class FilterSetFilterForFieldTests(TestCase):
 class FilterSetFilterForReverseFieldTests(TestCase):
 
     def test_reverse_o2o_relationship(self):
-        f = Account._meta.get_field_by_name('profile')[0]
+        f = Account._meta.get_field('profile')
         result = FilterSet.filter_for_reverse_field(f, 'profile')
         self.assertIsInstance(result, ModelChoiceFilter)
         self.assertEqual(result.name, 'profile')
@@ -192,7 +197,7 @@ class FilterSetFilterForReverseFieldTests(TestCase):
         self.assertEqual(result.extra['queryset'].model, Profile)
 
     def test_reverse_fk_relationship(self):
-        f = User._meta.get_field_by_name('comments')[0]
+        f = User._meta.get_field('comments')
         result = FilterSet.filter_for_reverse_field(f, 'comments')
         self.assertIsInstance(result, ModelMultipleChoiceFilter)
         self.assertEqual(result.name, 'comments')
@@ -201,7 +206,7 @@ class FilterSetFilterForReverseFieldTests(TestCase):
         self.assertEqual(result.extra['queryset'].model, Comment)
 
     def test_reverse_m2m_relationship(self):
-        f = Book._meta.get_field_by_name('lovers')[0]
+        f = Book._meta.get_field('lovers')
         result = FilterSet.filter_for_reverse_field(f, 'lovers')
         self.assertIsInstance(result, ModelMultipleChoiceFilter)
         self.assertEqual(result.name, 'lovers')
@@ -210,7 +215,7 @@ class FilterSetFilterForReverseFieldTests(TestCase):
         self.assertEqual(result.extra['queryset'].model, User)
 
     def test_reverse_non_symmetrical_selfref_m2m_field(self):
-        f = DirectedNode._meta.get_field_by_name('inbound_nodes')[0]
+        f = DirectedNode._meta.get_field('inbound_nodes')
         result = FilterSet.filter_for_reverse_field(f, 'inbound_nodes')
         self.assertIsInstance(result, ModelMultipleChoiceFilter)
         self.assertEqual(result.name, 'inbound_nodes')
@@ -219,7 +224,7 @@ class FilterSetFilterForReverseFieldTests(TestCase):
         self.assertEqual(result.extra['queryset'].model, DirectedNode)
 
     def test_reverse_m2m_field_with_through_model(self):
-        f = Worker._meta.get_field_by_name('employers')[0]
+        f = Worker._meta.get_field('employers')
         result = FilterSet.filter_for_reverse_field(f, 'employers')
         self.assertIsInstance(result, ModelMultipleChoiceFilter)
         self.assertEqual(result.name, 'employers')
@@ -558,19 +563,29 @@ class FilterSetOrderingTests(TestCase):
         self.assertQuerysetEqual(
             f.qs, ['carl', 'alex', 'jacob', 'aaron'], lambda o: o.username)
 
-    @unittest.skip('todo')
     def test_ordering_uses_filter_name(self):
         class F(FilterSet):
             account = CharFilter(name='username')
-
             class Meta:
                 model = User
                 fields = ['account', 'status']
                 order_by = True
 
-        f = F({'o': 'username'}, queryset=self.qs)
+        f = F({'o': 'account'}, queryset=self.qs)
         self.assertQuerysetEqual(
             f.qs, ['aaron', 'alex', 'carl', 'jacob'], lambda o: o.username)
+
+    def test_reverted_ordering_uses_filter_name(self):
+        class F(FilterSet):
+            account = CharFilter(name='username')
+            class Meta:
+                model = User
+                fields = ['account', 'status']
+                order_by = True
+
+        f = F({'o': '-account'}, queryset=self.qs)
+        self.assertQuerysetEqual(
+            f.qs, ['jacob', 'carl', 'alex', 'aaron'], lambda o: o.username)
 
     def test_ordering_with_overridden_field_name(self):
         """
@@ -642,14 +657,14 @@ class FilterSetTogetherTests(TestCase):
         self.alex = User.objects.create(username='alex', status=1)
         self.jacob = User.objects.create(username='jacob', status=2)
         self.qs = User.objects.all().order_by('id')
-    
+
     def test_fields_set(self):
         class F(FilterSet):
             class Meta:
                 model = User
                 fields = ['username', 'status', 'is_active', 'first_name']
                 together = [
-                    ('username', 'status'), 
+                    ('username', 'status'),
                     ('first_name', 'is_active'),
                 ]
 
@@ -660,14 +675,14 @@ class FilterSetTogetherTests(TestCase):
         f = F({'username': 'alex', 'status': 1}, queryset=self.qs)
         self.assertEqual(f.qs.count(), 1)
         self.assertQuerysetEqual(f.qs, [self.alex.pk], lambda o: o.pk)
-        
+
     def test_single_fields_set(self):
         class F(FilterSet):
             class Meta:
                 model = User
                 fields = ['username', 'status']
                 together = ['username', 'status']
-        
+
         f = F({}, queryset=self.qs)
         self.assertEqual(f.qs.count(), 2)
         f = F({'username': 'alex'}, queryset=self.qs)