+++ tests/test_widgets.py
@@ -4,6 +4,7 @@ from __future__ import unicode_literals
 from django.test import TestCase
 from django.forms import TextInput, Select
 
+from django_filters.widgets import BooleanWidget
 from django_filters.widgets import RangeWidget
 from django_filters.widgets import LinkWidget
 from django_filters.widgets import LookupTypeWidget
@@ -135,3 +136,30 @@ class RangeWidgetTests(TestCase):
             -
             <input type="text" name="price_1" value="9.99" />""")
 
+    def test_widget_attributes(self):
+        w = RangeWidget(attrs={'type': 'date'})
+        self.assertEqual(len(w.widgets), 2)
+        self.assertHTMLEqual(w.render('date', ''), """
+            <input type="date" name="date_0" />
+            -
+            <input type="date" name="date_1" />""")
+
+
+class BooleanWidgetTests(TestCase):
+    """
+    """
+    def test_widget_value_from_datadict(self):
+        """
+        """
+        w = BooleanWidget()
+
+        trueActive = {'active': 'true'}
+        result = w.value_from_datadict(trueActive, {}, 'active')
+        self.assertEqual(result, True)
+
+        falseActive = {'active': 'false'}
+        result = w.value_from_datadict(falseActive, {}, 'active')
+        self.assertEqual(result, False)
+
+        result = w.value_from_datadict({}, {}, 'active')
+        self.assertEqual(result, None)